Trello Task 254: Move Core/Base/MenuItem.php to Core/Lib/MenuItem.php - Renamed some missing RolAccess to RoleAccess according to previous changes.
