+++ Core/Base/MenuManager.php
@@ -19,6 +19,7 @@
 
 namespace FacturaScripts\Core\Base;
 
+use FacturaScripts\Core\Lib\MenuItem;
 use FacturaScripts\Core\Model;
 
 /**
@@ -103,7 +104,7 @@ class MenuManager
     private function pageNeedSave($pageModel, $pageData)
     {
         return (
-            ($pageModel->menu !== $pageData['menu']) || ($pageModel->submenu !== $pageData['submenu']) || 
+            ($pageModel->menu !== $pageData['menu']) || ($pageModel->submenu !== $pageData['submenu']) ||
             ($pageModel->title !== $pageData['title']) || ($pageModel->icon !== $pageData['icon']) ||
             ($pageModel->showonmenu !== $pageData['showonmenu'])
         );
@@ -165,7 +166,7 @@ class MenuManager
             if ($menuItem->name === $pageModel->name) {
                 $menu[$key]->active = true;
                 break;
-            } else if(!empty($pageModel->submenu) && !empty($menuItem->menu) && $menuItem->name === $pageModel->submenu) {
+            } elseif (!empty($pageModel->submenu) && !empty($menuItem->menu) && $menuItem->name === $pageModel->submenu) {
                 $menu[$key]->active = true;
                 $this->setActiveMenuItem($menu[$key]->menu, $pageModel);
                 break;
@@ -269,7 +270,7 @@ class MenuManager
      *
      * @param string $nick
      *
-     * @return Model\RolAccess[]
+     * @return Model\RoleAccess[]
      */
     private function getUserAccess($nick)
     {
@@ -277,7 +278,7 @@ class MenuManager
         $roleUserModel = new Model\RoleUser();
         $filter = [new DataBase\DataBaseWhere('nick', $nick)];
         foreach ($roleUserModel->all($filter) as $roleUser) {
-            foreach($roleUser->getRoleAccess() as $roleAccess) {
+            foreach ($roleUser->getRoleAccess() as $roleAccess) {
                 $access[] = $roleAccess;
             }
         }
