+++ Core/XMLView/ListRoleAccess.xml
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  *
- * Initial description for the controller ListRolAccess
+ * Initial description for the controller ListRoleAccess
  *
  * @author Artex Trading sa <jcuello@artextrading.com>
 -->
